requirements.txt
django
gunicorn
django-heroku
whitenoise
djangorestframework